+++ system/src/apps/_module.mjs
@@ -8,6 +8,7 @@ export {default as LightSourceTrackerSD} from "./LightSourceTrackerSD.mjs";
 export {default as LoadingSD} from "./LoadingSD.mjs";
 export {default as MonsterImporterSD} from "./MonsterImporterSD.mjs";
 export {default as NpcAttackRangesSD} from "./NpcAttackRangesSD.mjs";
+export {default as RequestCheckSD} from "./RequestCheckSD.mjs";
 export {default as ShadowdarklingImporterSD} from "./ShadowdarklingImporterSD.mjs";
 export {default as SpellBookSD} from "./SpellBookSD.mjs";
 export {default as SpellImporterSD} from "./SpellImporterSD.mjs";