+++ system/src/utils/UtilitySD.mjs
@@ -6,8 +6,23 @@ export default class UtilitySD {
 		return game.permissions.ACTOR_CREATE.includes(game.user.role);
 	}
 
-	/* Create a roll Macro from an Item dropped on the hotbar.
-	 * Get an existing item macro if one exists, otherwise create a new one.
+
+	static combineCollection(map1, map2) {
+		map2.forEach(value => {
+			if (map1.has(value._id)) {
+				shadowdark.warn(`Map already contains an item with key ${key}`);
+			}
+			else {
+				map1.set(value._id, value);
+			}
+		});
+
+		return map1;
+	}
+
+
+	/** Create a roll Macro from an Item dropped on the hotbar.
+	 *  Get an existing item macro if one exists, otherwise create a new one.
 	 *
 	 * @param {object} data - The dropped data
 	 * @param {number} slot - The hotbar slot to use
@@ -54,10 +69,46 @@ export default class UtilitySD {
 		game.user.assignHotbarMacro(macro, slot);
 	}
 
+
+	static diceSound() {
+		const sounds = [CONFIG.sounds.dice];
+		const src = sounds[0];
+		game.audio.play(src, {volume: 1});
+	}
+
+
 	static foundryMinVersion(version) {
 		const majorVersion = parseInt(game.version.split(".")[0]);
 		return majorVersion >= version;
 	}
+
+
+	// Work out the current Actor.
+	// If the user is the GM then use the current token they have selected.
+	//
+	static async getCurrentActor() {
+		let actor = null;
+
+		if (game.user.isGM) {
+			const controlledTokenCount = canvas.tokens.controlled.length;
+			if (controlledTokenCount > 0) {
+				if (controlledTokenCount !== 1) {
+					return ui.notifications.warn(
+						game.i18n.localize("SHADOWDARK.error.too_many_tokens_selected")
+					);
+				}
+				else {
+					actor = canvas.tokens.controlled[0].actor;
+				}
+			}
+		}
+		else {
+			actor = game.user.character;
+		}
+
+		return actor;
+	}
+
 
 	/**
 	 * Creates de-duplicated lists of Selected and Unselected Items.
@@ -87,6 +138,7 @@ export default class UtilitySD {
 		return [selectedItems, unselectedItems];
 	}
 
+
 	static async getFromUuid(uuid) {
 		const itemObj = await fromUuid(uuid);
 		if (itemObj) {
@@ -97,6 +149,7 @@ export default class UtilitySD {
 		}
 	}
 
+
 	static getFromUuidSync(uuid) {
 		const itemObj =  fromUuidSync(uuid);
 		if (itemObj) {
@@ -106,6 +159,7 @@ export default class UtilitySD {
 			return {name: "[Invalid ID]", uuid: uuid};
 		}
 	}
+
 
 	static async getItemsFromRollResults(results) {
 		const items = [];
@@ -124,6 +178,7 @@ export default class UtilitySD {
 		return items;
 	}
 
+
 	static getMessageStyles() {
 		const messageStyles = this.foundryMinVersion(12)
 			? CONST.CHAT_MESSAGE_STYLES
@@ -131,6 +186,7 @@ export default class UtilitySD {
 
 		return messageStyles;
 	}
+
 
 	static getNextDieInList(die, allDice) {
 		if (die === false) return die;
@@ -144,140 +200,13 @@ export default class UtilitySD {
 		return die;
 	}
 
+
 	static async getSlugifiedItemList(items) {
 		const itemList = {};
 		items.map(i => itemList[i.name.slugify()] = i.name );
 		return itemList;
 	}
 
-	static combineCollection(map1, map2) {
-		map2.forEach(value => {
-			if (map1.has(value._id)) {
-				shadowdark.warn(`Map already contains an item with key ${key}`);
-			}
-			else {
-				map1.set(value._id, value);
-			}
-		});
-
-		return map1;
-	}
-
-	// If this is a new release, show the release notes to the GM the first time
-	// they login
-	static async showNewReleaseNotes() {
-		if (game.user.isGM) {
-			const savedVersion = game.settings.get("shadowdark", "systemVersion");
-			const systemVersion = game.system.version;
-
-			if (systemVersion !== savedVersion) {
-				Hotbar.toggleDocumentSheet(
-					CONFIG.SHADOWDARK.JOURNAL_UUIDS.RELEASE_NOTES
-				);
-
-				game.settings.set(
-					"shadowdark", "systemVersion",
-					systemVersion
-				);
-			}
-		}
-	}
-
-	static async sleep(millisecs=1000) {
-		return new Promise((resolve, reject) => {
-  			setTimeout(resolve, millisecs);
-		});
-	}
-
-	/**
-	 * Asks the user for input if necessary for an effect that requires said input.
-	 * @param {Item} item - Item that has the effects
-	 * @param {*} effect - The effect being analyzed
-	 * @param {*} key - Optional key if it isn't a unique system.bonuses.key
-	 * @returns {Object} - Object updated with the changes
-	 */
-	static async modifyEffectChangesWithInput(item, effect, key = false) {
-		// Create an object out of the item to modify before creating
-		const itemObject = item.toObject();
-		let name = itemObject.name;
-
-		const changes = await Promise.all(
-			effect.changes.map(async c => {
-				if (CONFIG.SHADOWDARK.EFFECT_ASK_INPUT.includes(c.key)) {
-					const effectKey = (key) ? key : c.key.split(".")[2];
-
-					// Ask for user input
-					let linkedName;
-					[c.value, linkedName] = await item._handlePredefinedEffect(
-						effectKey,
-						null,
-						name
-					);
-
-					if (c.value) {
-						name += ` (${linkedName})`;
-					}
-				}
-				return c;
-			})
-		);
-
-		// Modify the Effect object
-		itemObject.effects.map(e => {
-			if (e._id === effect._id) {
-				e.changes = changes;
-				itemObject.name = name;
-			}
-			return e;
-		});
-		return itemObject;
-	}
-
-	/**
-	 * Contains logic that handles any complex effects, where the user
-	 * needs to provide input to determine the effect.
-	 * @param {Item} item - The item being created
-	 */
-	static async createItemWithEffect(item) {
-		let itemObj = item.toObject();
-		await Promise.all(itemObj.effects?.map(async e => {
-
-			// If the item contains effects that require user input,
-			// ask and modify talent before creating
-			if (
-				e.changes?.some(c =>
-					CONFIG.SHADOWDARK.EFFECT_ASK_INPUT.includes(c.key)
-				)
-			) {
-				// Spell Advantage requires special handling as it uses the `advantage` bons
-				if (e.changes.some(c => c.key === "system.bonuses.advantage")) {
-					// If there is no value with REPLACME, it is another type of advantage talent
-					if (e.changes.some(c => c.value === "REPLACEME")) {
-						const key = "spellAdvantage";
-						itemObj = await this.modifyEffectChangesWithInput(item, e, key);
-					}
-				}
-				else {
-					itemObj = await this.modifyEffectChangesWithInput(item, e);
-				}
-			}
-		}));
-
-		// If any effects was created without a value, we don't create the item
-		if (itemObj.effects.some(e => e.changes.some(c => !c.value))) return ui.notifications.warn(
-			game.i18n.localize("SHADOWDARK.item.effect.warning.add_effect_without_value")
-		);
-
-		// Activate lightsource tracking
-		if (itemObj.effects.some(e => e.changes.some(c => c.key === "system.light.template"))) {
-			const duration = itemObj.totalDuration;
-			itemObj.system.light.isSource = true;
-			itemObj.system.light.longevitySecs = duration;
-			itemObj.system.light.remainingSecs = duration;
-			itemObj.system.light.longevityMins = duration / 60;
-		}
-		return itemObj;
-	}
 
 	static isPrimaryGM() {
 		if (!game.user.isGM) return false;
@@ -302,6 +231,51 @@ export default class UtilitySD {
 			}
 		}
 	}
+
+
+	static async loadLegacyArtMappings() {
+		// search modules for legacy art mappings and convert to new format
+		for (const module of game.modules) {
+			if (!module.active) continue;
+			const flags = module.flags?.[module.id];
+			if (flags?.["shadowdark-art"]) {
+				module.flags.compendiumArtMappings = {
+					shadowdark: {
+						mapping: flags["shadowdark-art"],
+					},
+				};
+			}
+		}
+	}
+
+
+	// If this is a new release, show the release notes to the GM the first time
+	// they login
+	static async showNewReleaseNotes() {
+		if (game.user.isGM) {
+			const savedVersion = game.settings.get("shadowdark", "systemVersion");
+			const systemVersion = game.system.version;
+
+			if (systemVersion !== savedVersion) {
+				Hotbar.toggleDocumentSheet(
+					CONFIG.SHADOWDARK.JOURNAL_UUIDS.RELEASE_NOTES
+				);
+
+				game.settings.set(
+					"shadowdark", "systemVersion",
+					systemVersion
+				);
+			}
+		}
+	}
+
+
+	static async sleep(millisecs=1000) {
+		return new Promise((resolve, reject) => {
+  			setTimeout(resolve, millisecs);
+		});
+	}
+
 
 	static async toggleItemDetails(target) {
 		const listObj = $(target).parent();
@@ -332,24 +306,4 @@ export default class UtilitySD {
 		listObj.toggleClass("expanded");
 	}
 
-	static diceSound() {
-		const sounds = [CONFIG.sounds.dice];
-		const src = sounds[0];
-		game.audio.play(src, {volume: 1});
-	}
-
-	static async loadLegacyArtMappings() {
-		// search modules for legacy art mappings and convert to new format
-		for (const module of game.modules) {
-			if (!module.active) continue;
-			const flags = module.flags?.[module.id];
-			if (flags?.["shadowdark-art"]) {
-				module.flags.compendiumArtMappings = {
-					shadowdark: {
-						mapping: flags["shadowdark-art"],
-					},
-				};
-			}
-		}
-	}
 }